+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.8.3](https://github.com/antfu/wenyan-lang-vscode/compare/v0.8.2...v0.8.3) (2019-12-22)
+
+
+### ⚡ Features
+
+* compiler version in status bar ([dbd30ba](https://github.com/antfu/wenyan-lang-vscode/commit/dbd30baf33059b3e102ecab5bfe81ed47c183988))
+
+
+### 🐞 Bug Fixes
+
+* using modal for prompt render result ([19a5083](https://github.com/antfu/wenyan-lang-vscode/commit/19a5083670889c84c23c61bed2d97a35b7a1df9c))
+
+### [0.8.2](https://github.com/antfu/wenyan-lang-vscode/compare/v0.8.1...v0.8.2) (2019-12-22)
+
+
+### 🐞 Bug Fixes
+
+* vendored compiler ([e0836e0](https://github.com/antfu/wenyan-lang-vscode/commit/e0836e09a342a90a6c829873415bc1f178e2c7cf))
+
+### [0.8.1](https://github.com/antfu/wenyan-lang-vscode/compare/v0.8.0...v0.8.1) (2019-12-22)
+
+
+### ⚡ Features
+
+* update syntax ([93de7b3](https://github.com/antfu/wenyan-lang-vscode/commit/93de7b34d221000cb6bcd846b94922ee078be952))
+
 ## [0.8.0](https://github.com/antfu/wenyan-lang-vscode/compare/v0.7.4...v0.8.0) (2019-12-21)